Remove theme-utils.js. Revert theme_controller and added subscribe to matchMedia event.
