+++ hugo/src/js/app.js
@@ -18,5 +18,3 @@ if (process.env.NODE_ENV === 'production' || process.env.MIX_TURBO) {
 
 require('./highlight');
 require('./icons');
-
-require('./theme-utils').detectTheme();
